[NEST-FE-52] change to component modal instead of static page
